refactor: Add "Resources" link to navbar component and create resources page
This commit adds a new link to the navbar component for the "Resources" page. It also creates a new page component called "resources" that displays a list of AI resources with their corresponding links. This enhancement improves the navigation of the website and provides a dedicated page for users to access AI resources easily.